requirements.txt
python-telegram-bot==20.3
feedparser==6.0.10
googletrans==4.0.0-rc1
requests==2.31.0
beautifulsoup4==4.12.2
